Update 20250107 - Turret Projectile Update
Implemented the Turret Projectile Entity base class for the Ballista Arrow, along with a new head pitch control for the turrets, allowing them to have a proper head pitch for their projectile's arc motion.

Further updates will be implemented to further refine the said base class and to clean up the codes of its subclasses.

[CHANGELOG]
🟢 Added documentation to nearly all the fields in Turret Projectile entity base class.
🟢 Added comments to provide summary of what's happening in the "onEntityHit" method.
🟢 Added GeoEntity interface method implementation for the Bullet Entity.
🟢 Added a new data to track for all Turret entities. This will allow them to correctly render their heads when shooting. This will allow them to pitch their heads on the correct angle to match their projectile's trajectory.
🟡 Updated Ballista Arrow's superclass from Persistent Projectile to Turret Projectile.
🔴 Removed unused imports in Ballista Arrow entity class.
🔴 Removed redundant methods in Ballista Arrow entity after changing its superclass.